plsync: Fix rotator control scheduling using float
The target index where the external rotator should be adjusted comes
from the multiplication between sps (a small number) and the absolute
sample index (a large uint64_t). The previous implementation was quickly
losing precision significantly due to using sps as a single-precision
floating-point number. To solve the problem, cast sps to a double in the
computation.

+++ lib/plsync_cc_impl.cc
@@ -192,7 +192,7 @@ void plsync_cc_impl::control_rotator_freq(uint64_t abs_sof_idx,
     // operates on the sample stream (i.e. on samples, not symbols). Use the
     // known oversampling ratio and the calibrated symbol-spaced tag delay to
     // schedule the phase increment update.
-    uint64_t target_idx = d_sps * (abs_next_sof_idx + d_rot_ctrl.tag_delay);
+    uint64_t target_idx = (double)d_sps * (abs_next_sof_idx + d_rot_ctrl.tag_delay);
 
     // Prevent two frequency corrections at the same sample offset. The scenario
     // where this becomes possible is as follows:
